+++ NEWS.md
@@ -1,3 +1,5 @@
 #bayesdistreg 1.1.0
   * New functions rwMHgen() and indepMHgen() added for generic random walk and independence Metropolis-Hastings that are generic, and are applicable to constrained parameter spaces.
+  * New function psimval() added
+  * Function quant_bdr() revised for performance
   * Edited documentation for improved clarity